bia_app/requirements.txt
Flask 
Flask-SQLAlchemy 
Flask-Migrate 
Flask-WTF 
Flask-Login 
Flask-Bcrypt 
email_validator
Flask-Moment
python-dotenv
pandas
dnspython>=2.6.1 # not directly required, pinned by Snyk to avoid a vulnerability
werkzeug>=3.0.6 # not directly required, pinned by Snyk to avoid a vulnerability
zipp>=3.19.1 # not directly required, pinned by Snyk to avoid a vulnerability
numpy>=1.22.2 # not directly required, pinned by Snyk to avoid a vulnerability
pyotp==2.8.0
gunicorn